requirements.txt
Flask
Flask-Cors
Flask-SocketIO
PyJWT
openai
pymongo
bcrypt
